Simplify resource code, somewhat
