+++ server/handlers/resource_handler.go
@@ -2,18 +2,13 @@ package handlers
 
 import (
 	"github.com/dmoles/adler/server/resources"
-	"github.com/dmoles/adler/server/util"
 	"github.com/gorilla/mux"
-	"io"
 	"log"
 	"net/http"
 	"path"
 	"strconv"
 	"strings"
 )
-
-// TODO: centralize resource utility code & also use for templates
-// TODO: in makefile: statik -src=resources
 
 type resourceHandler struct {
 	dir     string
@@ -30,26 +25,20 @@ func (h *resourceHandler) Handle(w http.ResponseWriter, r *http.Request) {
 	}
 	relativePathClean := path.Clean(relativePath)
 	resourcePath := path.Join(h.dir, relativePathClean)
-	file, err := resources.Open(resourcePath)
+
+	resource, err := resources.Get(resourcePath)
 	if err != nil {
 		http.NotFound(w, r)
 		return
 	}
 
-	fileInfo, err := file.Stat()
-	if err != nil {
-		http.NotFound(w, r)
-		return
-	}
-
-	contentType := util.ContentType(resourcePath)
+	contentType := resource.ContentType()
 	w.Header().Add("Content-Type", contentType)
 
-	size := fileInfo.Size()
-	contentLength := strconv.FormatInt(size, 10)
-	w.Header().Add("Content-Length", contentLength)
+	size := resource.Size()
+	w.Header().Add("Content-Length", strconv.FormatInt(size, 10))
 
-	n, err := io.Copy(w, file)
+	n, err := resource.Copy(w)
 	if err != nil {
 		log.Printf("Error serving %#v: %v", resourcePath, err)
 	}
