+++ server/resources/http_bundle.go
@@ -35,6 +35,7 @@ func (r *httpBundle) RelativePath(absPath string) string {
 	return strings.TrimPrefix(absPath, "/")
 }
 
+// TODO: cache these?
 func (r *httpBundle) Get(resourcePath string) (Resource, error) {
 	f, err := r.Open(resourcePath)
 	if err != nil {
