+++ server/templates/templates.go
@@ -3,10 +3,8 @@ package templates
 import (
 	"fmt"
 	"github.com/dmoles/adler/server/resources"
-	"io"
 	"log"
 	"path"
-	"strings"
 	"text/template"
 )
 
@@ -35,20 +33,21 @@ type PageData struct {
 // ------------------------------------------------------------
 // Unexported
 
+// TODO: cache these, we're not going to have many of them
 func load(name string) *template.Template {
 	tmplPath := path.Join("/templates", name)
-	tmplFile, err := resources.Open(tmplPath)
+	resource, err := resources.Get(tmplPath)
 	if err != nil {
 		msg := fmt.Sprintf("Error locating template %s: %v", tmplPath, err)
 		log.Fatal(msg)
 	}
-	sb := new(strings.Builder)
-	_, err = io.Copy(sb, tmplFile)
+
+	tmplData, err := resource.AsString()
 	if err != nil {
 		msg := fmt.Sprintf("Error reading template %s: %v", tmplPath, err)
 		log.Fatal(msg)
 	}
-	tmplData := sb.String()
+
 	tmpl, err := template.New(name).Parse(tmplData)
 	if err != nil {
 		msg := fmt.Sprintf("Error parsing template %s: %v", tmplPath, err)
